+++ test/test_templates.py
@@ -68,3 +68,19 @@ def test_pyproject_no_docs(cli_output_no_docs, snapshot):
         pyproject = tomllib.load(f)
     assert pyproject.get('dependency-groups', {}).get('docs') is None
     assert pyproject == snapshot
+
+
+@pytest.mark.parametrize(
+    'src_file',
+    [
+        '__init__.py',
+        'app.py',
+        'logging.py',
+    ],
+)
+def test_src_dir(cli_output, snapshot, src_file):
+    """Files in src/ are correct."""
+    project_path, cli_result = cli_output
+    with open(project_path / 'src' / 'transporter_logs' / src_file, 'r', encoding='utf-8') as f:
+        file = f.read()
+    assert file == snapshot